// src/modules/invoce/facade/invoice.facade.interface.ts
export interface InvoiceFacadeFindInputDTO{
    id: string;

}

export interface InvoiceFacadeFindOutputDTO{
    id: string;
  name: string;
  document: string;
  address: {
    street: string;
    number: string;
    complement: string;
    city: string;
    state: string;
    zipCode: string;
  };
  items: {
    id: string;
    name: string;
    price: number;
  }[];
  total: number;
  createdAt: Date;
}

export interface InvoiceFacadeGenerateInputDTO{
    id?:string
    name: string;
    document: string;
    street: string;
    number: string;
    complement: string;
    city: string;
    state: string;
    zipCode: string;
    items: {
      id: string;
      name: string;
      price: number;
    }[];
}

export interface InvoiceFacadeGenerateOutputDTO{
    id: string;
    name: string;
    document: string;
    street: string;
    number: string;
    complement: string;
    city: string;
    state: string;
    zipCode: string;
    items: {
      id: string;
      name: string;
      price: number;
    }[];
    total: number;
}

export default interface InvoiceFacadeInterface{
    find(input:InvoiceFacadeFindInputDTO): Promise<InvoiceFacadeFindOutputDTO>;
    generate(input:InvoiceFacadeGenerateInputDTO):Promise<InvoiceFacadeGenerateOutputDTO>;
}